Remove the click handlers from any element above input fields to fix cm
Some sites (e.g. login.live.com) don't use the HTML5 placeholder tag but
uses a custom solution for this (probably because IE < 9 don't support
this). By adding this script we remove any click handlers on these DOM elements
living in front of the input field. The click event is then called on
the input field instead of the overlapping elements.

Signed-off-by: Tobia De Koninck <tobia@ledfan.be>